+++ CHANGELOG.md
@@ -5,6 +5,9 @@
       `null`, `isClass(Number)` to be true for numeric literals and `isClass(String)` to be true for string literals.
     * Added `isLiteral(value)` to check for a literal value
 - Fixed a bug where all unknown types were reported as `Boolean`
+- Added support for `undefined` type
+- Added a number of helpers for checking for literal types to `ReflectedTypeRef`
+
 # v0.0.18
 
 - Added support for type literal types, ie `foo(bar : false, baz : null, foobar : 123)`